Worked on things. Renamed app to Mobile OCR on the phone. tts_init welcomes user to app and tells them to tap and hold for instructions.
Tried to add a tap and launch new activity but it doesn't work because of the static reference can't be used on the intents start activity. ..... So I don't know what to do for that.

Started looking at the camera and how it works. Cleaned it up a little.

Made the screenreader launch by touching the MENU key since I can't get the tap to work.

Double checked the screenreader's lifecycle. It doesn't need modifications because it goes through a whole lifecycle. I did add an on resume to let the user know they are in the screenreader.