Changed the network notification so that the TTS tells you to connect rather than the vibrations (although either would be acceptable). Removed a bunch of unused id's. The OCR thread has been removed.
Something I just noticed, but should be on the TODO is that the onResume / onPause should be handled better. The TTS engine probably should be shut down and then restarted on resume. However, there are some issues with how then, the TTS should welcome the user (since there could be more than one initialization per onCreate of MobileOCR)